feat(page): add contact page
Closes #30
