+++ src/components/pages/contact/contact.css
@@ -1,0 +1,41 @@
+.banner {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -15%);
+  --name-size: 3vw;
+  --number-size: 5vw;
+  --email-size: 2vw;
+}
+
+.banner * {
+  margin: 0;
+  padding: 0;
+}
+
+.banner h1 {
+  font-family: Merriweather, serif;
+  font-size: var(--number-size);
+}
+
+.banner h2 {
+  font-family: Roboto, sans-serif;
+}
+
+.name {
+  font-size: var(--name-size);
+}
+
+.email {
+  text-align: right;
+  font-size: var(--email-size);
+}
+
+@media only screen and (max-width: 900px) {
+  .banner {
+    width: 90vw;
+    --name-size: 5vw;
+    --number-size: 7vw;
+    --email-size: 4vw;
+  }
+}
