+++ src/components/pages/home/home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Nav from "../Nav/Nav";
-import "./page.css";
+import Nav from "../../Nav/Nav";
+import "../page.css";
 import "./home.css";
 
 export default function Home({ classInjection }) {
